fixed #1

+++ src/Main.java
@@ -2,8 +2,8 @@ public class Main {
     public static void main(String[] args) {
         int weight = 70;
         int percent = 100 * 100;
-        int height = 193 * 193;
-        int coefficient = weight * percent / height;
+        float height = 193 * 193;
+        float coefficient = weight * percent / height;
         System.out.println(coefficient);
     }
 }
